[utils] dont download consul if exists

+++ consul-utils/install-consul.sh
@@ -45,9 +45,14 @@ echo "Creting user and group $CONSUL_USER..."
 useradd -q $CONSUL_USER
 
 # download, unarchive and copy consul to proper path
-echo 'Downloading and copying Consul...'
-wget "https://releases.hashicorp.com/consul/0.7.0/$CONSUL_ARCHIVE"
-mv $CONSUL_ARCHIVE ~/
+if [ ! -f ~/$CONSUL_ARCHIVE ];
+then
+	echo 'Downloading and copying Consul...'
+	wget "https://releases.hashicorp.com/consul/0.7.0/$CONSUL_ARCHIVE"
+	mv $CONSUL_ARCHIVE ~/
+else
+	echo "Not downloading Consul; file exists: $CONSUL_ARCHIVE"
+fi
 unzip ~/$CONSUL_ARCHIVE -d $CONSUL_PATH
 set_permissions "$CONSUL_PATH/consul" $CONSUL_USER $CONSUL_USER '500'
